requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
pydantic==2.9.2
pypdf2==3.0.1
python-dotenv==1.0.0
colorlog==6.8.2
pytest==8.3.3
streamlit==1.31.0
